refactor: @Size같은 기존의 애너테이션 포맷과 통일
