+++ src/main/java/gift/validation/ContainsOnlyAllowedSpecialCharacterValidator.java
@@ -1,6 +1,5 @@
-package gift.validator;
+package gift.validation;
 
-import gift.customAnnotation.ContainsOnlyAllowedSpecialCharacter;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
@@ -10,7 +9,7 @@ import java.util.Set;
 
 public class ContainsOnlyAllowedSpecialCharacterValidator implements ConstraintValidator<ContainsOnlyAllowedSpecialCharacter, String> {
 
-    private static final Set<Character> ALLOWED_SPECIAL_CHARACTERS = new HashSet<>(Arrays.asList('(', ')', '[', ']', '+', '-', '&', '/', '_',' '));
+    private static final Set<Character> ALLOWED_SPECIAL_CHARACTERS = new HashSet<>(Arrays.asList('(', ')', '[', ']', '+', '-', '&', '/', '_', ' '));
 
     @Override
     public boolean isValid(String value, ConstraintValidatorContext context) {
