+++ src/main/java/gift/validation/NotContainsKakao.java
@@ -1,6 +1,5 @@
-package gift.customAnnotation;
+package gift.validation;
 
-import gift.validator.NotContainsKakaoValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -14,7 +13,7 @@ import java.lang.annotation.Target;
 @Constraint(validatedBy = NotContainsKakaoValidator.class)
 public @interface NotContainsKakao {
 
-    String message() default "`카카오`가 포함된 문구는 담당 MD와 협의한 경우에만 사용 가능합니다.";
+    String message() default "{gift.validation.NotContainsKakao.message}";
 
     Class<?>[] groups() default {};
